+++ include/libc.h
@@ -38,6 +38,6 @@ int setpriority(unsigned int pid,unsigned int priority);
 
 void provoca_PageFault();
 
-int printbox(int x, int y, int ample, int alcada, char *missatge);
+int printbox(Byte x, Byte y, int ample, int alcada, char *missatge);
 
 #endif  /* __LIBC_H__ */
